Remove `CutEngine2DPieceInputData` class entirely
The `CutEngine2DPieceInputData` class has been completely removed from the `CutEngine2DPieceInputData.cs` file. This class was part of the `CutCraftEngineWebSocketCGLService.DataInput` namespace and was responsible for handling piece input data for a 2D cutting engine. It included properties such as width, height, count, rotatability, and ID, as well as methods for calculating surplus, real size, and determining if a piece is rotatable.