Add CGLConsoleLogger for logging CutEngine execution results
Introduced a new class `CGLConsoleLogger` in the `CutCraftEngineWebSocketCGLService.CGLCalculator` namespace to log results to the console when `CutEngine.Execute` is called. Imported necessary namespaces to support this functionality. The class includes a private readonly `_calculator` field and a constructor that subscribes to the `InitializeExecutedEvent` of `CGLCalculator`. The `Display` method is invoked upon event trigger, utilizing `CGLConsoleResultPrinter`.