Refactor CGLLayoutFactory to use factory pattern
Updated the constructor of the `CGLLayoutFactory` class to accept `ICGLDataOutputFactory<CGLSheet>` and `ICGLDataOutputFactory<CGLPart>` instead of `List<CGLSheet>` and `List<CGLPart>`. The `_sheets` and `_parts` fields are now initialized by calling the `Get()` method on the provided factory instances. This change introduces a factory pattern, enhancing flexibility and decoupling object creation from the `CGLLayoutFactory` class.